pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coinglass-mcp"
version = "0.1.0"
description = "CoinGlass derivatives data MCP server (funding rates, OI, long/short, liquidations)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.23.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0", "pytest-asyncio>=0.23", "respx>=0.21"]

[project.scripts]
coinglass-mcp = "coinglass_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/coinglass_mcp"]
